requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.1.0
langsmith>=0.4.0
weaviate-client>=3.26.7,<4.0.0

# PDF processing
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5
pandas>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0

# Optional: For enhanced table extraction
tabula-py>=2.9.0
camelot-py[cv]>=0.11.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0